==> 03_treatment_text/01_create_dataframe_from_xml.ipynb <==
# 1. Crear dataframe por especialidad - 
### Directorio: df_specialty_title_abstract
### Columnas dataframe: (id, title_spa, abstract_spa) 

import os
import xml.etree.ElementTree as ET
import pandas as pd

path_pubmed_xml_specialties = '../02_download_pubmed/specialties_case_report_xml/'
path_df_specialty_title_abstract = './dataframes/df_specialty_title_abstract_case_report'

def get_lists_pmid_title_abstract(root):
    
    list_pmid = []
    list_title = []
    list_abstract = []
    
    for PubmedArticle in root.findall('PubmedArticle'):
        
        pmid = ''
        title_spa = ''
        abstract_spa = ''
        
        for MedlineCitation in PubmedArticle.findall('MedlineCitation'):
            pmid = MedlineCitation.find('PMID').text
        
            for Article in MedlineCitation.findall('Article'):
                if not Article.find('VernacularTitle') is None:              
                    title_spa = Article.find('VernacularTitle').text
                    
                    if not title_spa is None and len(title_spa) > 1 and title_spa.isupper():
                        title_spa = title_spa.replace("'A", "Á")
                        title_spa = title_spa.replace("'E", "É")
                        title_spa = title_spa.replace("'I", "Í")
                        title_spa = title_spa.replace("'O", "Ó")
                        title_spa = title_spa.replace("'U", "Ú")
                        
            
            for OtherAbstract in MedlineCitation.findall('OtherAbstract'):
                abstrac_lang = OtherAbstract.get('Language')
                                
                if abstrac_lang == 'spa':
                    for AbstractText in OtherAbstract.findall('AbstractText'):
                        if not AbstractText.text is None:
                            abstract_spa = abstract_spa + AbstractText.text + ' '
        
        if not title_spa is None and len(title_spa) > 1: 
            list_pmid.append(pmid)
            list_title.append(title_spa)
            list_abstract.append(abstract_spa)
        
    print(len(list_pmid), len(list_title), len(list_abstract))  
    
    return list_pmid, list_title, list_abstract

def read_xml_file(file_xml):
    tree = ET.parse(file_xml)
    root = tree.getroot()
    return root

def create_dataframe_text_abstract(specialty_name, specialty_xml):
    
    xml_data = read_xml_file(specialty_xml)
    
    #pmid = get_pmid(xml_data)
    list_pmid, list_title, list_abstract = get_lists_pmid_title_abstract(xml_data)
    
    df = pd.DataFrame ( { 'id': list_pmid, 'title': list_title, 'abstract': list_abstract})
    df.to_csv(os.path.join(path_df_specialty_title_abstract, specialty_name) + '.csv')

    print("Guardado ", os.path.join(path_df_specialty_title_abstract, specialty_name) + '.csv' , ' Long: ', len(df))


def read_path_specialties_xml():
    
    for root, dirs, list_files in os.walk(path_pubmed_xml_specialties):
        for specialty in list_files:
            specialty_name = specialty.split(".xml")[0]
            
            print(specialty_name)

            create_dataframe_text_abstract(specialty_name, root + specialty)
            

read_path_specialties_xml()

# 3. Crear dataframe con seis columnas
### Directorio: df_specialty_ngram
### Columnas dataframe: (idpubmed, título-unigramas, título-bigramas, título-trigramas, abstract-unigramas, abstract-bigramas, abstract-trigramas)

## Procesamiento de texto:
    1 Lowercase: pasar a minúscula si el término no es uppercase
    2. Unigramas: 
        2.1 borrar puntuacion
        2.2 borrar dígitos

import os
import pandas as pd
import os
import spacy
import ast
nlp = spacy.load('es_core_news_sm')
from nltk import everygrams
from nltk.corpus import stopwords
stop_words = set(stopwords.words('spanish')) 
import string

path_df_specialty_title_abstract = './pubmed_files/dataframes/df_specialty_title_abstract'
path_df_specialty_xgram_title_abstract = './pubmed_files/dataframes/df_specialty_ngram'

def change_to_lowercase(term):
    if not term.isupper():
        return term.lower()
    return term

def to_lowercase(list_terms):
    
    list_terms_new = []
    
    for term in list_terms:
        
        if type(term) == str:
             list_terms_new.append(change_to_lowercase(term))
                
        elif type(term) == tuple:
            new_tuple = ()
            
            for t in term:
                new_tuple = new_tuple + (change_to_lowercase(t),)
            list_terms_new.append(new_tuple)  
    
    #print("2",list_terms_new )
    return list_terms_new

def remove_stopword_punt_digit(list_tems):
    
    list_tems = [term for term in list_tems if not term in stop_words]
    list_tems = [term for term in list_tems if not term in set(string.punctuation)]
    list_tems = [term for term in list_tems if not term.isdigit() ] 
    
    #print("3", list_tems )
    return list_tems

def tokenize(text):
    list_tokens = []
    
    if type(text) == float:
        text = ''
    
    doc = nlp(text)
    for token in doc:
        list_tokens.append(token.text)
        
    return list_tokens

def obtain_grams(text, gram):
    list_tokens = tokenize(text)
    list_bigrams = list(everygrams(list_tokens, min_len=gram, max_len=gram))
    
    return list_bigrams

def create_new_df_xgrams(df):
    
    list_id = []
    list_ngram = []
    
    # Iteración por filas del DataFrame:
    for index, row in df.iterrows():
        
        file_id = row['id']
        abstract = row['abstract']
        title = row['title']
        
        title_unig = tokenize(title)
        title_big = obtain_grams(title ,2)
        title_trig = obtain_grams(title ,3)
        
        abstract_unig = tokenize(abstract)
        abstract_big = obtain_grams(abstract ,2)
        abstract_trig = obtain_grams(abstract ,3)
       
        ## procesamiento
        title_unig = remove_stopword_punt_digit(to_lowercase(title_unig))
        title_big = to_lowercase(title_big)
        title_trig = to_lowercase(title_trig)

        abstract_unig = remove_stopword_punt_digit(to_lowercase(abstract_unig))
        abstract_big = to_lowercase(abstract_big)
        abstract_trig = to_lowercase(abstract_trig)
        
        len_total = len(title_unig) + len(title_big) + len(title_trig) + len(abstract_unig) + len(abstract_big) + len(abstract_trig)
        
        list_id.extend([file_id] * len_total)
        list_ngram.extend(title_unig)
        list_ngram.extend(title_big)
        list_ngram.extend(title_trig)
        
        list_ngram.extend(abstract_unig)
        list_ngram.extend(abstract_big)
        list_ngram.extend(abstract_trig)
    
    new_df = pd.DataFrame ( { 'id': list_id, 
                         'ngram': list_ngram 
                         })
    return new_df
    

def read_df_specialties_title_abstract():

    list_df = os.listdir(path_df_specialty_title_abstract)
    for specialty_df in list_df:
        
        specialty_df = os.path.join(path_df_specialty_title_abstract, specialty_df)
        specialty_name_csv = specialty_df.split("/")[-1]
        
        df = pd.read_csv(specialty_df)
        print("Specialty: ", specialty_name_csv , ' --- Longitud del DF:', len(df))
        
        new_df = create_new_df_xgrams(df)
        
        file_out = os.path.join(path_df_specialty_xgram_title_abstract, specialty_name_csv)
        new_df.to_csv(file_out)
        print("Guardado ", file_out, ' Long: ', len(new_df))

   


read_df_specialties_title_abstract()

# 4 Crear otro tipo de estructura

import os
import pandas as pd
import pickle

path_df_specialty_xgram_title_abstract = './pubmed_files/dataframes/df_specialty_ngram'

def read_df_specialties_grams():
    
    dic_final = {}

    list_df = os.listdir(path_df_specialty_xgram_title_abstract)
    
    for specialty_df in list_df:
        
        dic_specialty = {}
        
        specialty_name_csv = specialty_df
        
        df = pd.read_csv(os.path.join(path_df_specialty_xgram_title_abstract, specialty_df))
        
        print("Specialty: ", specialty_name_csv , ' --- Longitud del DF:', len(df))
        
        list_docs = []
        dic_terms = {}
        
        # Iteración por filas del DataFrame:
        for index, row in df.iterrows():

            file_id = row['id']
            ngram = row['ngram']
            
            list_docs.append(file_id)
            dic_terms.setdefault(ngram, []).append(file_id)
            
        dic_specialty['terms'] = dic_terms
        dic_specialty['docs'] = list(set(list_docs))
        
        print("Nº de términos en la especialidad:", len(dic_specialty['terms']))
        print("Nº de doc en la especialidad:", len(dic_specialty['docs']))
        dic_final[specialty_name_csv] = dic_specialty
               
    with open('dic_specialties.pkl', 'wb') as f:
        pickle.dump(dic_final, f)    

read_df_specialties_grams()

### Prueba varios términos en especialidad:

def deserialize_object(path):
    pickle_in = open(path,"rb")
    obj = pickle.load(pickle_in)
    pickle_in.close()
    print("Cargado el objeto", path.split("/")[- 1])
    return obj

path_dic = 'dic_specialties.pkl'
dic = deserialize_object(path_dic)

print(dic['vaccinology.csv']['terms']["('de', 'la')"]) # 19 veces -> correcto
print(dic['vaccinology.csv']['terms']["vacuna"]) # 13 veces -> correcto



print(dic['vaccinology.csv']['docs'])
